+++ packages/w3wallets/src/metamask/metamask.ts
@@ -22,6 +22,7 @@ export class Metamask extends Wallet {
 
     //////
     await this.page.getByTestId("onboarding-import-wallet").click();
+    await this.page.getByTestId("onboarding-import-with-srp-button").click();
     await this.page
       .getByTestId("srp-input-import__srp-note")
       .pressSequentially(mnemonic);
@@ -96,44 +97,57 @@ export class Metamask extends Wallet {
     return text;
   }
 
-  async connectToNetwork(networkName: string): Promise<void>;
-  async connectToNetwork(settings: NetworkSettings): Promise<void>;
-  async connectToNetwork(settingsOrName: NetworkSettings | string) {
-    if (typeof settingsOrName !== "string") {
-      await this.page.locator(".mm-picker-network").click();
-      await this.page
-        .getByRole("button", { name: "Add a custom network" })
-        .click();
-      await this.page
-        .getByTestId("network-form-network-name")
-        .fill(settingsOrName.name);
-      await this.page
-        .getByTestId("network-form-chain-id")
-        .fill(settingsOrName.chainId.toString());
-      await this.page
-        .getByTestId("network-form-ticker-input")
-        .fill(settingsOrName.currencySymbol);
+  async connectToNetwork(
+    networkName: string,
+    networkType: "Default" | "Custom" = "Default",
+  ) {
+    await this.page.getByTestId("sort-by-networks").click();
+    await this.page.getByRole("button", { name: networkType }).click();
 
-      await this.page.getByTestId("test-add-rpc-drop-down").click();
-      await this.page.getByRole("button", { name: "Add RPC URL" }).click();
-      await this.page
-        .getByTestId("rpc-url-input-test")
-        .fill(settingsOrName.rpc);
-      await this.page.getByRole("button", { name: "Add URL" }).click();
-      await this.page.getByRole("button", { name: "Save" }).click();
+    const additionalNetwork = this.page
+      .getByTestId("additional-network-item")
+      .getByText(networkName);
+
+    try {
+      await additionalNetwork.isEnabled({ timeout: 1000 });
+      await additionalNetwork.click();
+      await this.page.getByTestId("confirmation-submit-button").click();
+    } catch (error) {
+      await this.page.getByText(networkName).click();
+      await this.page.getByTestId("modal-header-close-button").click();
     }
+  }
 
-    await this.page.locator(".mm-picker-network").click();
+  async addCustomNetwork(settings: NetworkSettings) {
+    await this.page.getByTestId("account-options-menu-button").click();
+    await this.page.getByTestId("global-menu-networks").click();
+    await this.page
+      .getByRole("button", { name: "Add a custom network" })
+      .click();
+    await this.page
+      .getByTestId("network-form-network-name")
+      .fill(settings.name);
+    await this.page
+      .getByTestId("network-form-chain-id")
+      .fill(settings.chainId.toString());
+    await this.page
+      .getByTestId("network-form-ticker-input")
+      .fill(settings.currencySymbol);
+
+    await this.page.getByTestId("test-add-rpc-drop-down").click();
+    await this.page.getByRole("button", { name: "Add RPC URL" }).click();
+    await this.page.getByTestId("rpc-url-input-test").fill(settings.rpc);
+    await this.page.getByRole("button", { name: "Add URL" }).click();
+    await this.page.getByRole("button", { name: "Save" }).click();
+  }
+
+  async enableTestNetworks() {
+    await this.page.getByTestId("account-options-menu-button").click();
+    await this.page.getByTestId("global-menu-networks").click();
     await this.page
       .locator("text=Show test networks >> xpath=following-sibling::label")
       .click();
-    await this.page
-      .getByTestId(
-        typeof settingsOrName === "string"
-          ? settingsOrName
-          : settingsOrName.name,
-      )
-      .click();
+    await this.page.keyboard.press("Escape");
   }
 
   async approve() {